+++ CHANGELOG.md
@@ -44,8 +44,12 @@
 * Produce Collector can brush Armadillos for Armadillo Scutes on 1.21+
 * Auto-Brewer can brew Oozing, Weaving, Wind Charging and Infestation potions
 * Auto crafters now require the new Crafter block in their recipes
-* Add recipes for new 1.21 blocks and armor
 * Wolf Armor craftable in the Armor Forge from Armadillo Scutes
+* Copper Bulb craftable in the Enhanced Crafting Table
+* Ominous Bottle craftable in the Magic Workbench
+* Trial Key craftable in the Magic Workbench
+* Vault and Trial Spawner craftable in the Enhanced Crafting Table
+* Ore Crusher and Electric Ore Grinder can crush Tuff and its variants into Sand
 
 * Butcher Android collects Breeze Rods, Wind Charges and mushrooms from new mobs
 * Butcher Android collects Creaking Hearts from Creakings